requirements.txt

cbor~=1.0.0
PyYAML~=5.3
python-can~=3.3.2
paho-mqtt~=1.5.0
